+++ client/src/features/relatedLists/RelatedListAccordionItem.tsx
@@ -8,7 +8,7 @@ import { IHalLink } from '../../types/IHalLink'
 import StyledHr from '../../styles/shared/Hr'
 import theme from '../../styles/theme'
 
-import SemanticRelatedList from './SemanticRelatedList'
+import RelatedList from './RelatedList'
 
 interface IProps {
   searchTermConfig: IHalLink
@@ -72,7 +72,7 @@ const RelatedListAccordionItem: React.FC<IProps> = ({
     <React.Fragment>
       <div
         className="accordion-item"
-        data-testid={`semantic-list-accordion-item-${searchTerm}`}
+        data-testid={`related-list-accordion-item-${searchTerm}`}
       >
         <h2
           className="accordion-header"
@@ -89,11 +89,11 @@ const RelatedListAccordionItem: React.FC<IProps> = ({
               pushSiteImproveEvent(
                 'Accordion Item',
                 activeAccordion ? 'Close' : 'Open',
-                title,
+                `Accordion ${title}`,
               )
               setActiveAccordion(!activeAccordion)
             }}
-            data-testid={`semantic-list-accordion-item-${searchTerm}-button`}
+            data-testid={`related-list-accordion-item-${searchTerm}-button`}
           >
             {isLoading ? 'Loading...' : title}
           </StyledAccordionButton>
@@ -106,7 +106,7 @@ const RelatedListAccordionItem: React.FC<IProps> = ({
           <div className="accordion-body">
             {/* Render list of results against facets API endpoint */}
             {activeAccordion && isSuccess && data && data.results !== null && (
-              <SemanticRelatedList
+              <RelatedList
                 results={data.results}
                 halLink={halLink}
                 next={data.next}