+++ client/CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 - Added SI event for pagination ([#168](https://github.com/project-lux/lux-frontend/issues/168)).
 - Added SI event for search switch ([#162](https://github.com/project-lux/lux-frontend/issues/162)).
 - Added SI event for entity links ([#164](https://github.com/project-lux/lux-frontend/issues/164)).
+- Added SI event for accordions ([#166](https://github.com/project-lux/lux-frontend/issues/166)).
 - Added SI event for feedback button on entity pages ([#91](https://github.com/project-lux/lux-frontend/issues/91)).
 - Added SI event for search links ([#90](https://github.com/project-lux/lux-frontend/issues/90)).